Move test_multi_global into a new file.
